🤘🏼Conectar categorias, eliminar buscador sin uso
